docs(fix): Updated ER diagram
Changes made:
For the **author rating**, the name can be: author_score
- We should not store author_phone because it is PII (Personal Identification Information) and storing it is not secure.
- We should include the no_of_books column because it is a standard practice, and we should have such a column instead of a derived column.

**Book Table**
-  For the book_rating column, we can set the default value as 0(this mean that the book is not yet been purchased)
-  We should include the columns written in the optional section except book_cover_image because
"Instead, what is commonly done is to store references or URLs to the location of the images in the database. For each book record in your database, you might have a field that contains the URL or path pointing to the location of the corresponding book image. When users access the online bookstore, the application retrieves these URLs from the database and uses them to display the images on the user interface." -- source: chatgpt.

**Premium_type table**
- subscription_status instead of premium_type, and all the column names will be replaced by "subscription".
- I don't think we should include the "premium image" column because it is not needed, but we do need the premium_duration column, and for this we can add the default values, like a customer can take a premium for 3M, 6M or 12M.
-  For the subscription_name we will all  think of some quirky names.

**Customer_table**
- What does primary_customer_phone mean?
- According to me, we should add another column for 2 phone numbers.
-  And if we decide to add the premium_duration, then we can add the premium_expire_date, or we can also derive it.

Now, according to me, we should add one wishlist column because it is very basic and will look like the following:

**Wishlist:**
wishlist_id (primary key)
customer_id (foreign key to Customers table)
book_id (foreign key to Books table)
date_added

And if we decide to add the above table, then we also need to do the following column in the customer table:

wishlist (foreign key to Wishlist table)

**Customer_address**
-  Yes we do need this table

**Shopping Cart**
- According to me, the columns mentioned in the optional section might help us run some analysis. I know we won't be running any analysis, but reviewers can ask us.

**Average_rating**
- Yes, we need this table because it will help on the business side and will help the owner take action in cases where ratings are low.

- default date for people who do not have a birthday (search on google)
- For authentication and authorization, there is a separate table for the user (in this table, user with rights are listed)
- visit the following website for project name suggestions, and by tomorrow everyone will select 3 names, and by tomorrow we will decide on a name: https://namify.tech/suggestions

Decide on the following
- wish list feature
- ⁠magazines and different categories features
- ⁠decide on premium types for ayon

TODO:
- suggest subscription names
- auto increment id's
- format before commit
- on delete cascade and check ';'

Reviewed by: @GauravBohra2001 @amanpreet-5045 @vishveshrathore20 @kratikasharma01
References: other DB Models